python3 versions and initial ucf101 files

+++ tflib/save_images.py
@@ -16,7 +16,7 @@ def save_images(X, save_path):
     while n_samples % rows != 0:
         rows -= 1
 
-    nh, nw = rows, n_samples/rows
+    nh, nw = rows, (int)(n_samples/rows)
 
     if X.ndim == 2:
         X = np.reshape(X, (X.shape[0], int(np.sqrt(X.shape[1])), int(np.sqrt(X.shape[1]))))
@@ -28,10 +28,10 @@ def save_images(X, save_path):
         img = np.zeros((h*nh, w*nw, 3))
     elif X.ndim == 3:
         h, w = X[0].shape[:2]
-        img = np.zeros((h*nh, w*nw))
+        img = np.zeros((h*nh,w*nw))
 
     for n, x in enumerate(X):
-        j = n/nw
+        j = (int)(n/nw)
         i = n%nw
         img[j*h:j*h+h, i*w:i*w+w] = x
